ControlPCbyEmail/requirements.txt
pillow
email